test(taos): move sql, stmt and stmt2 perf tests to tests
