+++ taos/tests/sql_ws_interlace.rs
@@ -7,8 +7,8 @@ use chrono::Local;
 use rand::Rng;
 use taos::{AsyncQueryable, AsyncTBuilder, TaosBuilder};
 
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
+#[tokio::test]
+async fn test_sql_ws_interlace() -> anyhow::Result<()> {
     let subtable_cnt = 100_0000;
     let data_cnt = 1_0000_0000;
 
@@ -25,12 +25,14 @@ async fn main() -> anyhow::Result<()> {
         &format!("create database {db} vgroups 10"),
         &format!("use {db}"),
         "create stable s0 (ts timestamp, c1 int, c2 float, c3 float) tags(t1 int)",
+        // "create stable s0 (ts timestamp, c1 bool, c2 tinyint, c3 smallint, c4 int, c5 bigint,
+        //     c6 double, c7 binary(20)) tags(t1 int)",
     ])
     .await?;
 
     create_subtables(db, subtable_cnt).await;
 
-    let sqls = generate_sqls(subtable_cnt).await;
+    let sqls = generate_sqls(data_cnt, subtable_cnt).await;
 
     exec_sqls(db, sqls).await;
 
@@ -90,41 +92,73 @@ async fn create_subtables(db: &str, subtable_cnt: usize) {
     println!("create subtables done, elapsed = {:?}\n", start.elapsed());
 }
 
-async fn generate_sqls(subtable_cnt: usize) -> Vec<String> {
+async fn generate_sqls(data_cnt: usize, subtable_cnt: usize) -> Vec<String> {
     println!("generate sqls start");
 
     let start = Instant::now();
+
     let batch_cnt = 1_0000;
+    let sql_cnt = data_cnt / batch_cnt;
+
     let thread_cnt = 10;
-    let thread_subtable_cnt = subtable_cnt / thread_cnt;
+    let thread_sql_cnt = sql_cnt / thread_cnt;
+    let thread_loop_cnt = data_cnt / thread_cnt / subtable_cnt;
+
     let mut tasks = Vec::with_capacity(thread_cnt);
 
-    for i in (0..subtable_cnt).step_by(thread_subtable_cnt) {
+    for i in 0..thread_cnt {
         let task = tokio::spawn(async move {
-            let mut sqls = Vec::with_capacity(thread_subtable_cnt);
+            println!("thread[{i}] generate sqls start");
+
+            let start = Instant::now();
             let mut rng = rand::thread_rng();
+            let mut sqls = Vec::with_capacity(thread_sql_cnt);
 
-            for j in 0..thread_subtable_cnt {
-                let mut sql = String::with_capacity(100 * batch_cnt);
-                sql.push_str(&format!("insert into d{} values ", i + j));
-
+            for _ in 0..thread_loop_cnt {
                 let ts = SystemTime::now()
                     .duration_since(UNIX_EPOCH)
                     .unwrap()
                     .as_millis() as i64;
 
-                for k in 0..batch_cnt {
-                    let ts = ts + k as i64;
-                    let c1: i32 = rng.gen();
-                    let c2: f32 = rng.gen_range(0.0..1000_0000.);
-                    let c2: f32 = (c2 * 100.0).round() / 100.0;
-                    let c3: f32 = rng.gen_range(0.0..1000_0000.);
-                    let c3: f32 = (c3 * 100.0).round() / 100.0;
-                    sql.push_str(&format!("({ts},{c1},{c2},{c3}),"));
-                }
+                for k in (0..subtable_cnt).step_by(batch_cnt) {
+                    // insert into d0 values() d1 values() ...
+                    let mut sql = String::with_capacity(100 * batch_cnt);
+                    sql.push_str("insert into ");
+                    for l in 0..batch_cnt {
+                        let tbname = format!("d{}", k + l);
 
-                sqls.push(sql);
+                        let c1 = rng.gen::<i32>();
+
+                        let c2: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c2 = (c2 * 100.0).round() / 100.0;
+
+                        let c3: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c3 = (c3 * 100.0).round() / 100.0;
+
+                        sql.push_str(&format!("{tbname} values({ts},{c1},{c2},{c3}) "));
+
+                        // let c1 = rng.gen::<bool>() as u8;
+                        // let c2 = rng.gen::<i8>();
+                        // let c3 = rng.gen::<i16>();
+                        // let c4 = rng.gen::<i32>();
+                        // let c5 = rng.gen::<i64>();
+                        // let c6: f64 = rng.gen_range(0.0..1000_0000.);
+                        // let c6 = (c6 * 100.0).round() / 100.0;
+                        // let c7: String = (0..20).map(|_| rng.gen_range('a'..='z')).collect();
+
+                        // sql.push_str(&format!(
+                        //     "{tbname} values(now,{c1},{c2},{c3},{c4},{c5},{c6},'{c7}') "
+                        // ));
+                    }
+
+                    sqls.push(sql);
+                }
             }
+
+            println!(
+                "thread[{i}] generate sqls done, elapsed = {:?}",
+                start.elapsed()
+            );
 
             sqls
         });
@@ -132,7 +166,7 @@ async fn generate_sqls(subtable_cnt: usize) -> Vec<String> {
         tasks.push(task);
     }
 
-    let mut sqls = Vec::with_capacity(subtable_cnt);
+    let mut sqls = Vec::with_capacity(sql_cnt);
     for task in tasks {
         sqls.extend(task.await.unwrap());
     }
@@ -144,18 +178,19 @@ async fn generate_sqls(subtable_cnt: usize) -> Vec<String> {
 }
 
 async fn exec_sqls(db: &str, sqls: Vec<String>) {
-    println!("exec sqls start, start = {:?}", Local::now());
+    let now = Local::now();
+    println!("exec sqls start, start = {:?}", now);
 
     let start = Instant::now();
 
     let thread_cnt = 4;
-    let thread_sql_cnt = sqls.len() / thread_cnt;
+    let chunk_size = sqls.len() / thread_cnt;
 
     let mut tasks = vec![];
 
-    for (i, sqls) in sqls.chunks(thread_sql_cnt).enumerate() {
-        let sqls = sqls.to_vec();
-
+    for (i, chunk) in sqls.chunks(chunk_size).enumerate() {
+        let chunk = chunk.to_vec();
+        // let taos = taos.clone();
         let taos = TaosBuilder::from_dsn("ws://localhost:6041")
             .unwrap()
             .build()
@@ -167,7 +202,7 @@ async fn exec_sqls(db: &str, sqls: Vec<String>) {
         let task = tokio::spawn(async move {
             println!("thread[{i}] exec sqls start");
             let start = Instant::now();
-            for sql in sqls {
+            for sql in chunk {
                 if let Err(err) = taos.exec(sql).await {
                     eprintln!("thread[{i}] exec sql failed, err = {err:?}");
                 }
