+++ taos/tests/stmt2_interlace.rs
@@ -1,4 +1,7 @@
-use std::time::{Instant, SystemTime, UNIX_EPOCH};
+use std::{
+    os::unix::thread,
+    time::{Instant, SystemTime, UNIX_EPOCH},
+};
 
 use chrono::Local;
 use flume::{Receiver, Sender};
@@ -6,8 +9,8 @@ use rand::Rng;
 use taos::{AsyncQueryable, AsyncTBuilder, ColumnView, Stmt2, TaosBuilder};
 use taos_query::stmt2::{AsyncBindable, Stmt2BindData};
 
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
+#[tokio::test]
+async fn test_stmt2_interlace() -> anyhow::Result<()> {
     let subtable_cnt = 100_0000;
     let data_cnt = 1_0000_0000;
 
@@ -18,7 +21,7 @@ async fn main() -> anyhow::Result<()> {
         .build()
         .await?;
 
-    let db = "db_202412161921";
+    let db = "db_202412121942";
 
     taos.exec_many([
         &format!("drop database if exists {db}"),
@@ -99,7 +102,11 @@ async fn create_subtables(db: &str, subtable_cnt: usize) {
     println!("create subtables done, elapsed = {:?}\n", start.elapsed());
 }
 
-async fn generate_datas(txs: Vec<Sender<Stmt2BindData>>, data_cnt: usize, subtable_cnt: usize) {
+async fn generate_datas(
+    txs: Vec<Sender<Vec<Stmt2BindData>>>,
+    data_cnt: usize,
+    subtable_cnt: usize,
+) {
     let now = Local::now();
     println!("generate datas start, start = {now}");
 
@@ -108,58 +115,64 @@ async fn generate_datas(txs: Vec<Sender<Stmt2BindData>>, data_cnt: usize, subtab
     let batch_cnt = 1_0000;
     // let bind_datas_cnt = data_cnt / batch_cnt;
 
-    let thread_cnt = txs.len();
-    let thread_subtable_cnt = subtable_cnt / thread_cnt;
+    let thread_cnt = 4;
     // let thread_bind_data_cnt = bind_datas_cnt / thread_cnt;
     // let thread_loop_cnt = data_cnt / thread_cnt / subtable_cnt;
 
     // let mut tasks = Vec::with_capacity(thread_cnt);
 
-    for i in (0..subtable_cnt).step_by(thread_subtable_cnt) {
-        let sender = txs[i].clone();
+    for i in 0..thread_cnt {
+        let sender = txs[i % txs.len()].clone();
         tokio::spawn(async move {
+            let sender = sender;
             println!("thread[{i}] generate datas start");
+            let start = Instant::now();
 
             let mut rng = rand::thread_rng();
-            // 10000 * 10000 = 100000000
-            for j in 0..thread_subtable_cnt {
+            // let mut bind_datas = Vec::with_capacity(thread_bind_data_cnt);
+
+            for _ in 0..25 {
                 let ts = SystemTime::now()
                     .duration_since(UNIX_EPOCH)
                     .unwrap()
                     .as_millis() as i64;
 
-                // for k in (0..subtable_cnt).step_by(thread_subtable_cnt) {
-                let mut tss = Vec::with_capacity(batch_cnt);
-                let mut c1s = Vec::with_capacity(batch_cnt);
-                let mut c2s = Vec::with_capacity(batch_cnt);
-                let mut c3s = Vec::with_capacity(batch_cnt);
+                for k in (0..subtable_cnt).step_by(batch_cnt) {
+                    let mut datas = Vec::with_capacity(batch_cnt);
+                    for l in 0..batch_cnt {
+                        let c1 = rng.gen::<i32>();
+                        let c2: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c2 = (c2 * 100.0).round() / 100.0;
+                        let c3: f32 = rng.gen_range(0.0..1000_0000.);
+                        let c3 = (c3 * 100.0).round() / 100.0;
 
-                for l in 0..batch_cnt {
-                    let ts = ts + l as i64;
-                    let c1 = rng.gen::<i32>();
-                    let c2: f32 = rng.gen_range(0.0..1000_0000.);
-                    let c2 = (c2 * 100.0).round() / 100.0;
-                    let c3: f32 = rng.gen_range(0.0..1000_0000.);
-                    let c3 = (c3 * 100.0).round() / 100.0;
+                        let cols = vec![
+                            ColumnView::from_millis_timestamp(vec![ts]),
+                            ColumnView::from_ints(vec![c1]),
+                            ColumnView::from_floats(vec![c2]),
+                            ColumnView::from_floats(vec![c3]),
+                        ];
 
-                    tss.push(ts);
-                    c1s.push(c1);
-                    c2s.push(c2);
-                    c3s.push(c3);
+                        let tbname = format!("d{}", k + l);
+                        // println!("tbname = {tbname}");
+                        let data = Stmt2BindData::new(Some(tbname), None, Some(cols));
+                        datas.push(data);
+                    }
+
+                    // bind_datas.push(datas);
+                    sender.send(datas).unwrap();
                 }
-
-                let tbname = format!("d{}", i + j);
-                let cols = vec![
-                    ColumnView::from_millis_timestamp(tss),
-                    ColumnView::from_ints(c1s),
-                    ColumnView::from_floats(c2s),
-                    ColumnView::from_floats(c3s),
-                ];
-                let data = Stmt2BindData::new(Some(tbname), None, Some(cols));
-                sender.send(data).unwrap();
-                // }
             }
+
+            println!(
+                "thread[{i}] generate datas done, elapsed = {:?}",
+                start.elapsed()
+            );
+
+            // bind_datas
         });
+
+        // tasks.push(task);
     }
 
     // let mut bind_datas = Vec::with_capacity(bind_datas_cnt);
@@ -173,7 +186,7 @@ async fn generate_datas(txs: Vec<Sender<Stmt2BindData>>, data_cnt: usize, subtab
     // bind_datas
 }
 
-async fn stmt2_exec(mut rxs: Vec<Receiver<Stmt2BindData>>, db: &str) {
+async fn stmt2_exec(mut rxs: Vec<Receiver<Vec<Stmt2BindData>>>, db: &str) {
     let now = Local::now();
     println!("stmt2 exec start, start = {now}");
     let start = Instant::now();
@@ -202,7 +215,7 @@ async fn stmt2_exec(mut rxs: Vec<Receiver<Stmt2BindData>>, db: &str) {
             println!("thread[{i}] stmt2 exec start");
 
             while let Ok(datas) = rx.recv() {
-                if let Err(err) = stmt2.bind(&[datas]).await {
+                if let Err(err) = stmt2.bind(&datas).await {
                     eprintln!("thread[{i}] bind data failed, err = {err:?}");
                 }
 
@@ -227,3 +240,65 @@ async fn stmt2_exec(mut rxs: Vec<Receiver<Stmt2BindData>>, db: &str) {
 
     println!("stmt2 exec done, elapsed = {:?}", start.elapsed());
 }
+
+// async fn stmt2_exec(rxs: Vec<Receiver<Vec<Stmt2BindData>>>, db: &str) -> anyhow::Result<()> {
+//     let now = Local::now();
+//     println!("stmt2 exec start, start = {now}");
+
+//     let start = Instant::now();
+
+//     let thread_cnt = 10;
+//     let chunk_size = datas.len() / thread_cnt;
+
+//     let mut tasks = vec![];
+
+//     for (i, chunk) in datas.chunks(chunk_size).enumerate() {
+//         let chunk = chunk.to_vec();
+//         let db = db.to_owned();
+
+//         let task = tokio::spawn(async move {
+//             let taos = TaosBuilder::from_dsn("ws://localhost:6041")
+//                 .unwrap()
+//                 .build()
+//                 .await
+//                 .unwrap();
+
+//             taos.exec(format!("use {db}")).await.unwrap();
+
+//             let start = Instant::now();
+//             println!("thread[{i}] stmt2 exec start");
+
+//             let mut stmt2 = Stmt2::init(&taos).await.unwrap();
+
+//             let sql = "insert into s0 (tbname, ts, c1, c2, c3) values(?, ?, ?, ?, ?)";
+//             stmt2.prepare(sql).await.unwrap();
+
+//             for data in chunk {
+//                 if let Err(err) = stmt2.bind(&data).await {
+//                     eprintln!("thread[{i}] bind data failed, err = {err:?}");
+//                 }
+
+//                 let affacted = stmt2.exec().await.unwrap();
+//                 assert_eq!(affacted, 10000);
+//             }
+
+//             let elapsed = start.elapsed();
+//             println!("thread[{i}] stmt2 exec done, elapsed = {elapsed:?}");
+//             elapsed
+//         });
+
+//         tasks.push(task);
+//     }
+
+//     let durations: Vec<_> = futures::future::join_all(tasks)
+//         .await
+//         .into_iter()
+//         .filter_map(|res| res.ok())
+//         .collect();
+
+//     let elapsed = durations.iter().max();
+//     println!("Stmt2 exec took {:?}", elapsed.unwrap());
+//     println!("stmt2 exec done, elapsed = {:?}", start.elapsed());
+
+//     Ok(())
+// }
